Add maple_super_reduce command (requires Maple)
Maple has two functions to find Moser-irreducible form of the linear
differential system: super_reduce and moser_reduce. The former is
implemented in this commit. It is not exactly what we need since
as applied to git_410 system it introduces additional singular
points in oo. It may be still helpful to debug and test our algorithms.