Improve Algorithm 3
Now lee_81, git_409, and two examples from test_normalize.py are
normalized correctly. I still did not manage to normalize git_410 since
its Fuchsian form is too complicated and Maxima raises an error when
reading it.

The implementation is rather verbose with verbose output for
debugging which may affect the overal performance. It make sense to
optimize some hot points in the future, but for now stability and
correctness have priority.