Workaround problems in Maxima's 'solve'
On some expressions 'solve' fails to find roots of an expression,
even though it can easily factor it. Passing already factored
expression in seems to help for now.

Also 'solve' sometimes returns an expression like this as a root:

  1/33420*sqrt(1671)*(29*sqrt(1671) - 557*I)

... while other times it would return this for the same root:

  -1/60*I*sqrt(1671) + 29/20

Because of this, we need to use 'expand' to normalize roots.

I wonder if it's possible for 'solve' to return both of these forms
for the same expression. If it is, our 'singularities' will get broken.